Update code folder to match tested project

+++ code/lcd.h
@@ -1,6 +1,6 @@
 
-void lcd_show_height_percentage(short height_percentage);
+void lcd_show_height_percentage(int height_percentage);
 
-void lcd_show_reference_percentage(short reference_percentage);
+void lcd_show_reference_percentage(int reference_percentage);
 
-void lcd_show_control_percentage(short control_percentage);
+void lcd_show_control_percentage(int control_percentage);
